requirements.txt
requests
dnspython
colorama
tqdm
